[refactor]디테일페이지 유튜브 컴포넌트 리팩토링 (박진양)
